requirements.txt
Django>=5.0.7,<5.1
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.6,<2.10
faker>=18.13.0,<18.14
drf-spectacular>=0.26.3,<0.27
djangorestframework-simplejwt>=5.2.2,<5.3
django-cors-headers>=4.2.0,<4.3
requests>=2.31.0,<2.32
django-environ>=0.10.0,<0.11
Pillow>=10.0.0,<10.1
redis>=5.0.0,<5.1
celery>=5.3.4,<5.4
django-celery-beat==2.6.0
django-celery-results>=2.5.1,<2.6
flower>=2.0.1,<2.1
SQLAlchemy>=2.0.20,<2.1
google-api-python-client>=2.100.0,<2.101
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.0
django-mptt==0.16.0
django-filter==24.2
